+++ src/slitflow/fig/line.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 
-from .figure import Figure
+from .figure import Figure, inherit_split_depth
 from ..fun.misc import reduce_list as rl
 
 
@@ -19,7 +19,6 @@ class Simple(Figure):
             containing column names "img_no" and "trj_no", you have to set
             group_depth=2. If you set group_depth=1, all trajectories are
             misconnected to each other.
-        param["split_depth"] (int): File split depth number.
 
     Returns:
         Figure: Line Figure object
@@ -29,7 +28,7 @@ class Simple(Figure):
         """Copy info from reqs[0] and add params.
         """
         self.info.copy_req(0)
-        self.info.delete_column(keeps=self.info.get_column_name("index"))
+        inherit_split_depth(self, 0, param["group_depth"])
         self.info.add_param(
             "calc_cols", param["calc_cols"], "str", "X and Y columns")
         if "err_col" in param:
@@ -39,8 +38,6 @@ class Simple(Figure):
                 param["cap_size"] = 2
             self.info.add_param(
                 "cap_size", param["cap_size"], "float", "Error bar cap size")
-        self.info.set_group_depth(param["group_depth"])
-        self.info.set_split_depth(param["split_depth"])
 
     @staticmethod
     def process(reqs, param):
@@ -98,7 +95,6 @@ class WithModel(Figure):
             Required if "err_col" in param. Defaults to 2.
         param["group_depth"] (int): Data split depth number.
         param["group_depth_model"] (int): Depth number to split model data.
-        param["split_depth"] (int): File split depth number.
 
     Returns:
         Figure: Line Figure object with error and model
@@ -106,7 +102,7 @@ class WithModel(Figure):
 
     def set_info(self, param={}):
         self.info.copy_req(0)
-        self.info.delete_column(keeps=self.info.get_column_name("index"))
+        inherit_split_depth(self, 0, param["group_depth"])
         self.info.add_param(
             "calc_cols", param["calc_cols"], "str", "Columns to calculate")
         if "err_col" in param:
@@ -124,7 +120,6 @@ class WithModel(Figure):
             "index_cols_model", index_cols_model, "list of str",
             "Index column names of model curve")
         self.info.set_group_depth(param["group_depth"])
-        self.info.set_split_depth(param["split_depth"])
 
     @staticmethod
     def process(reqs, param):